requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
colored==2.2.4
cryptography==44.0.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mysql-connector-python==9.1.0
paramiko==3.5.0
propcache==0.2.1
psycopg2==2.9.10
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
tqdm==4.67.1
urllib3==2.3.0
walk==0.3.5
wheel==0.45.1
yarl==1.18.3
